备份

+++ src/main/java/top/cellargalaxy/dachuangspringboot/dataSet/DataSetSplit.java
@@ -4,5 +4,5 @@ package top.cellargalaxy.dachuangspringboot.dataSet;
  * Created by cellargalaxy on 17-9-28.
  */
 public interface DataSetSplit {
-	DataSet[] splitDataSet(DataSet dataSet, double testPro, double trainMissPro, double testMissPro, double trainLabel1Pro, double testLabel1Pro, int k);
+	DataSet[] splitDataSet(DataSet dataSet, double testPro, double trainMissPro, double testMissPro, double trainLabel1Pro, double testLabel1Pro, double k);
 }
